Actions: build custom docker image for bakery CI
Build a custom Ubuntu 24.10 image with PKCS11 Keyavault token. This is
required because GitHUb runners only offer Ubuntu 24.04, which has
systemd 255. We use PKCS11 signing feature of systemd-repart, which is
only avalable since systemd 256 (in Ubuntu 24.10).

The other reason is, we use PKCS11 Azure Keyvault token. It takes a
non-trivial amount of time to compile and install it, which would have
to be performed during every sysext build. Using a Docker image with a
prebuild token significantly speeds-up the build process.